Rename astar_search to astar_route
